schemaexpr: remove expr_filter.go
This commit removes the `pkg/sql/catalog/schemaexpr/expr_filter.go`
file. The `schemaexpr` package is intended to contain logic for dealing
with expressions defined in a schema, such as check constraints,
computed columns, and partial index predicates. The two exported
functions in the file, `RemapIVarsInTypedExpr` and `RunFilter` did not
fit this theme. They have been moved elsewhere.

Release note: None
